Small change
Deleted a bit of extraneous code that fixed a compile error on my end
(which ended up fixing itself) but had no real function. Hardware and
software both work well in real world testing. There might be a bit of
lag in the button handling, but I'm not sure. I haven't had much
opportunity to test the analog sticks, but they seem to work ok.